requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
openai==1.37.1
langgraph==0.2.14
pydantic==2.8.2
PyPDF2==3.0.1
itsdangerous==2.2.0
Werkzeug==3.0.4
Jinja2==3.1.4
Flask-SocketIO==5.3.6
pymongo==4.8.0
reportlab==4.2.2
httpx==0.27.2
Markdown==3.6
